RISC-V PLIC: merge drivers
merge the PLIC drivers for ariane and hifive
