+++ src/plat/ariane/config.cmake
@@ -16,7 +16,7 @@ if(KernelPlatformAriane)
     list(APPEND KernelDTSList "src/plat/ariane/overlay-ariane.dts")
     declare_default_headers(
         TIMER_FREQUENCY 10000000llu PLIC_MAX_NUM_INT 30
-        INTERRUPT_CONTROLLER drivers/irq/ariane.h
+        INTERRUPT_CONTROLLER drivers/irq/riscv-plic.h
     )
 else()
     unset(KernelPlatformFirstHartID CACHE)
